chore(config): update config files
- add .gitignore patterns for local files
- update TypeScript and Next.js configs for .mdx files
- (nit) prevent 'prettier/prettier' errors from erroring
