+++ eslint.config.mjs
@@ -21,6 +21,7 @@ const eslintConfig = [
       'react/no-unescaped-entities': 'off',
       '@typescript-eslint/no-unused-vars': 'warn',
       '@typescript-eslint/no-explicit-any': 'warn',
+      'prettier/prettier': 'warn',
     },
   }),
 ]
