--- next.config.mjs
@@ -1,14 +1,0 @@
-import createMDX from '@next/mdx'
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'md', 'mdx'],
-  turbopack: {},
-}
-
-const withMDX = createMDX({
-  extension: /\.mdx?$/,
-})
-
-export default withMDX(nextConfig)
