+++ next.config.ts
@@ -1,0 +1,20 @@
+import type { NextConfig } from 'next'
+import createMDX from '@next/mdx'
+
+const nextConfig: NextConfig = {
+  reactStrictMode: true,
+  pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'md', 'mdx'],
+  turbopack: {},
+  // Note: Using the Rust compiler means we cannot use
+  // rehype or remark plugins. If you need them, remove
+  // the `experimental.mdxRs` flag.
+  experimental: {
+    mdxRs: true,
+  },
+}
+
+const withMDX = createMDX({
+  extension: /\.mdx?$/,
+})
+
+export default withMDX(nextConfig)
